pyproject.toml
[project]
name = "kafka_emulator"
version = "0.0.1"
description = "Kafka Messages Emulator"
authors = [
    {name = "Siak Hooi",email = "siakhooi@gmail.com"}
]
keywords = [ "kafka" ]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "License :: OSI Approved :: MIT License",
    "Topic :: Utilities"
]

license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "kafka-python>=2.0.2",
    "pyyaml>=6.0",
]

[project.urls]
homepage = "https://github.com/siakhooi/kafka-emulator"
repository = "https://github.com/siakhooi/kafka-emulator"
documentation = "https://github.com/siakhooi/kafka-emulator/wiki"
"Bug Tracker" = "https://github.com/siakhooi/kafka-emulator/issues"

[project.scripts]
kafka-emulator = 'kafka_emulator.cli:run'


[tool.poetry]
packages = [{include = "kafka_emulator", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-cov = "^7.1.0"
flake8 = "^7.2.0"
pytest-mock = "^3.15.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
